Infra/ remove utils_old
