+++ eslint-rules/lib/utils/generalUtils.js
@@ -1,17 +1,19 @@
-function findValueNodeOfIdentifier(identifierName, context) {
-  const scope = context.getScope();
-  const {variables} = scope;
+function findValueNodeOfIdentifier(identifierName, scope) {
+  const varsInScope = scope.variables;
   let valueNode = false;
-  variables.forEach(variable => {
+  varsInScope.forEach((variable) => {
     if (variable.name === identifierName) {
       if (variable.defs) {
         valueNode = variable.defs[variable.defs.length - 1].node.init;
       }
     }
   });
-
-  return valueNode;
+  if (scope.upper === null) {
+    return valueNode;
+  }
+  return valueNode || findValueNodeOfIdentifier(identifierName, scope.upper);
 }
+
 
 module.exports = {
   findValueNodeOfIdentifier
